requirements.txt
streamlit
pandas
google-generativeai
ebisu
pydantic
python-dotenv
requests
